requirements.txt
crewai[tools]
playwright
pytest
python-dotenv
rich
pyyaml
jinja2
